requirements.txt
Django==3.1
django-rest-swagger==2.2.0
djangorestframework==3.12
pillow=>9.2